Rewrite error reporter in TypeScript
